refactor: move landed costs logic to different composable
